vec3 Package Revert
Looking to move back to the stable build of gl-matrix, with the hopes of using tree-shaking where possible; I will look further into whether the modern version of that package supports it yet. It has better type definitions than the split-up package used before, and I think I'd rather hope with trying to stay with the main package rather than a bundling-shim focused one. If the main package can't be tree shaken though, then that is possibly not going to work. It really shouldn't make the package that much more huge, ideally. Bundlephobia helps display this nicely. If it doesn't support tree shaking, then I can look into migrating that project over as well! Nice to work on things like that.